refactor: move test docs to fixture files, clean conftest
