+++ tests/integration/fixtures/docs/api-perf.md
@@ -1,0 +1,5 @@
+# API Performance
+
+Cache responses with Redis. Use connection pooling for upstream services.
+Rate limit with token bucket algorithm to prevent abuse. Enable gzip
+compression for large payloads. Use async I/O for non-blocking requests.
