+++ tests/integration/fixtures/docs/auth-part1.md
@@ -1,0 +1,5 @@
+# Authentication: OAuth Setup
+
+Configure OAuth 2.0 with client ID and secret. Register your application
+in the developer portal to obtain credentials. The authorization endpoint
+handles the initial redirect flow with PKCE challenge.
