+++ tests/integration/fixtures/docs/auth-part3.md
@@ -1,0 +1,5 @@
+# Authentication: Session Management
+
+Sessions stored in Redis with 24h TTL. Session IDs are generated
+using cryptographically secure random bytes. Inactive sessions
+are garbage collected every 6 hours by the cleanup worker.
