+++ tests/integration/fixtures/docs/notes.md
@@ -1,0 +1,5 @@
+# Meeting Notes March 2026
+
+Decided to migrate to PostgreSQL 16 for improved JSON support.
+Timeline: Q2 2026. Migration lead: Sarah Chen. Budget approved
+for dedicated DBA contractor. Rollback plan documented in wiki.
