+++ tests/integration/fixtures/docs/specs.md
@@ -1,0 +1,7 @@
+# Thunderbolt X500
+
+Engine: 3.5L V6 TurboForce
+Oil capacity: 6.5 quarts
+Top speed: 155 mph
+Horsepower: 365 hp
+Transmission: 8-speed automatic
